requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohappyeyeballs==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.5.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
class-registry==2.1.2
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
exceptiongroup==1.2.1
executor==23.2
fasteners==0.19
filters==1.3.2
fpdf==1.7.2
frozenlist==1.4.1
google-api-core==2.21.0
google-api-python-client==2.92.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
loguru==0.7.2
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.30.3
pdfkit==1.0.0
pillow==10.3.0
property-manager==3.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.2.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
regex==2024.7.24
register==0.1
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
state==0.1.1.dev0
tqdm==4.66.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.20
utils==1.0.2
verboselogs==1.7
wkhtmltopdf==0.2
yarl==1.9.4
